megatron-energon/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "megatron-energon"
# dynamic = ["version"]
version = "0.0.0"
authors = [
  { name="Lukas Vögtle", email="lvoegtle@nvidia.com" },
  { name="Philipp Fischer", email="pfischer@nvidia.com" },
]
description = "Megatron's multi-modal data loader"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]
dependencies = [
  "braceexpand",
  "click",
  "numpy",
  "pillow>=10.0.1",  # WEBP vulnerability fixed starting from 10.0.1
  "pyyaml",
  "torch",
  "tqdm",
  "webdataset",
  "s3fs",
]

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "sphinxcontrib-napoleon",
  "sphinx",
  "myst-parser",
  "sphinx-rtd-theme",
  "sphinx-click",
]
transforms = [
  "torchvision",  # Needed for megatron.energon.transforms
]

[project.urls]
Homepage = "https://github.com/NVIDIA/Megatron-Energon"

[tool.hatch.build.targets.wheel]
packages = ["src/megatron"]

[tool.hatch.build.targets.sdist]
packages = ["src/megatron"]

[project.scripts]
energon = "megatron.energon.cli.main:main"

[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.vscode
    | \.idea
    | _build
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
